Tests: move worlds loader to own folder
